fix: split flow component into two files
